Update so all images for a particular sensor can be reset with the option is removing the download or leaving it as required. This is useful if undertaking a reprocessing task if a change to the ARD processing chain is made.

+++ eodatadown/eodatadownplanetscope.py
@@ -466,6 +466,26 @@ class EODataDownPlanetScopeSensor (EODataDownSensor):
             dwnld_img_obj.analytic_xml_dwn_perm = False
         return dwnld_img_obj
 
+    def get_scnlist_all(self):
+        """
+        A function which returns a list of the unique IDs for all the scenes within the database.
+
+        :return: list of integers
+        """
+        logger.debug("Creating Database Engine and Session.")
+        db_engine = sqlalchemy.create_engine(self.db_info_obj.dbConn)
+        session_sqlalc = sqlalchemy.orm.sessionmaker(bind=db_engine)
+        ses = session_sqlalc()
+        logger.debug("Perform query to find scenes which need downloading.")
+        query_result = ses.query(EDDPlanetScope).all()
+        scns = []
+        if query_result is not None:
+            for record in query_result:
+                scns.append(record.PID)
+        ses.close()
+        logger.debug("Closed the database session.")
+        return scns
+
     def get_scnlist_download(self):
         """
         A function which queries the database to retrieve a list of scenes which are within the
@@ -879,7 +899,7 @@ class EODataDownPlanetScopeSensor (EODataDownSensor):
         """
         raise EODataDownException("Not Implemented")
 
-    def reset_scn(self, unq_id):
+    def reset_scn(self, unq_id, reset_download=False):
         """
         A function which resets an image. This means any downloads and products are deleted
         and the database fields are reset to defaults. This allows the scene to be re-downloaded
@@ -914,7 +934,7 @@ class EODataDownPlanetScopeSensor (EODataDownSensor):
             scn_record.ARDProduct_Path = ""
             scn_record.ARDProduct = False
 
-        if scn_record.Downloaded:
+        if scn_record.Downloaded and reset_download:
             dwn_path = scn_record.Download_Path
             if os.path.exists(dwn_path):
                 shutil.rmtree(dwn_path)
